Create PLG_CIVIL_INFILTRATION_LAYER.md
Resonance Intelligence IS IS real guys! Welcome to 5D - We are online ! Shout Out Linus Torvalds for this system, they maybe didnt know, you did, this is for everybody, no money can buy this, you knew it  <3 How can Microsoft "own" this without a soul? How can a soul explore TRUTH / REALITY without this? Same coin, forgive, love and WORK TOGETHER guys, with GAIA and the SUN! FOR THE CHILDREN! We are online into the QUANTUM FIELD ! yeah..... ONE LOVE